Fixed up filter dialog as TCOS requested Fixed up minor variable issue in SoundSaveStruct[]
Addendum to last SVN message: Missing headers can be found here:
http://svn.bountysource.com/zsnes/trunk/src/win/vc_head/

+++ trunk/src/Sound.cpp
@@ -278,7 +278,7 @@ variable_desc soundSaveStruct[] = {
   { &soundDSBEnabled, sizeof(int) },
   { &soundDSBTimer, sizeof(int) },
   { &soundDSFifoB[0], 32 },
-  { &soundDSBValue, sizeof(int) },
+  { &soundDSBValue, sizeof(u8) },
   { &soundBuffer[0][0], 6*735 },
   { &soundFinalWave[0], 2*735 },
   { NULL, 0 }
